docker: Also copy /lib64 and /usr/lib64 for Ubuntu images
Currently we only copy the contents of /usr/lib64 from RHEL images.
Ubuntu stores most libraries in architecture-specific libraries,
but /lib64 is still necessary as this is where the system expects
to find ld-linux.so:

    root@cac76d38073f:/# ls -l /lib64
    lrwxrwxrwx 1 root root 9 Oct  4 02:08 /lib64 -> usr/lib64

    root@cac76d38073f:/# ls -l /usr/lib64/
    lrwxrwxrwx 1 root root 42 Sep 25 14:45 ld-linux-x86-64.so.2 -> /lib/x86_64-linux-gnu/ld-linux-x86-64.so.2

Without this symlink, lld is still unable to link, although now
that we copy /lib/x86_64-linux-gnu it no longer complains about the
C libraries:

    % swift build --experimental-swift-sdk 5.9-RELEASE_ubuntu_jammy_x86_64
    Building for debugging...
    ...
    error: link command failed with exit code 1 (use -v to see invocation)
    ld.lld: error: /Users/euanh/Library/org.swift.swiftpm/swift-sdks/5.9-RELEASE_ubuntu_jammy_x86_64.artifactbundle/5.9-RELEASE_ubuntu_jammy_x86_64/x86_64-unknown-linux-gnu/ubuntu-jammy.sdk/lib/x86_64-linux-gnu/libc.so:5: cannot find /lib64/ld-linux-x86-64.so.2 inside /Users/euanh/Library/org.swift.swiftpm/swift-sdks/5.9-RELEASE_ubuntu_jammy_x86_64.artifactbundle/5.9-RELEASE_ubuntu_jammy_x86_64/x86_64-unknown-linux-gnu/ubuntu-jammy.sdk
    >>> GROUP ( /lib/x86_64-linux-gnu/libc.so.6 /usr/lib/x86_64-linux-gnu/libc_nonshared.a  AS_NEEDED ( /lib64/ld-linux-x86-64.so.2 ) )
    >>>                                                                                                 ^
    clang-13: error: linker command failed with exit code 1 (use -v to see invocation)
    error: fatalError
    [5/6] Linking hello-world

N.B.: On RHEL ld-linux.so is an ordinary file, not a symlink, and
/lib64 was always copied into the SDK:

    bash-5.1$ ls /lib64/ld-linux-x86-64.so.2
    /lib64/ld-linux-x86-64.so.2

+++ Sources/SwiftSDKGenerator/Generator/SwiftSDKGenerator+Copy.swift
@@ -50,16 +50,17 @@ extension SwiftSDKGenerator {
             '
             """#
           )
+        }
 
-          let sdkUsrLib64Path = sdkUsrPath.appending("lib64")
-          try await generator.copyFromDockerContainer(
-            id: containerID,
-            from: FilePath("/usr/lib64"),
-            to: sdkUsrLib64Path
-          )
+        let sdkUsrLib64Path = sdkUsrPath.appending("lib64")
+        try await generator.copyFromDockerContainer(
+          id: containerID,
+          from: FilePath("/usr/lib64"),
+          to: sdkUsrLib64Path
+        )
+        try await createSymlink(at: pathsConfiguration.sdkDirPath.appending("lib64"), pointingTo: "./usr/lib64")
 
-          try await createSymlink(at: pathsConfiguration.sdkDirPath.appending("lib64"), pointingTo: "./usr/lib64")
-
+        if case .rhel = self.versionsConfiguration.linuxDistribution {
           // `libc.so` is a linker script with absolute paths on RHEL, replace with a relative symlink
           let libcSO = sdkUsrLib64Path.appending("libc.so")
           try await removeFile(at: libcSO)
@@ -83,11 +84,11 @@ extension SwiftSDKGenerator {
             to: sdkUsrLibPath.appending(subpath)
           )
         }
+        try await generator.createSymlink(at: pathsConfiguration.sdkDirPath.appending("lib"), pointingTo: "usr/lib")
 
         // Python artifacts are redundant.
         try await generator.removeRecursively(at: sdkUsrLibPath.appending("python3.10"))
 
-        try await generator.createSymlink(at: pathsConfiguration.sdkDirPath.appending("lib"), pointingTo: "usr/lib")
         try await generator.removeRecursively(at: sdkUsrLibPath.appending("ssl"))
         try await generator.copyTargetSwift(from: sdkUsrLibPath)
         try await generator.stopDockerContainer(id: containerID)
